+++ member-service/src/main/java/com/hedvig/memberservice/external/trustpilot/TrustpilotReviewLinkRequestDto.kt
@@ -1,0 +1,8 @@
+package com.hedvig.memberservice.external.trustpilot
+
+data class TrustpilotReviewLinkRequestDto(
+    val referenceId: String,
+    val email: String,
+    val name: String,
+    val locale: String
+)
