+++ member-service/src/main/java/com/hedvig/memberservice/external/trustpilot/TrustpilotReviewLinkResponseDto.kt
@@ -1,0 +1,6 @@
+package com.hedvig.memberservice.external.trustpilot
+
+data class TrustpilotReviewLinkResponseDto(
+    val id: String,
+    val url: String
+)
