+++ member-service/src/main/java/com/hedvig/memberservice/services/trustpilot/TrustpilotReviewInvitation.kt
@@ -1,0 +1,6 @@
+package com.hedvig.memberservice.services.trustpilot
+
+data class TrustpilotReviewInvitation(
+    val id: String,
+    val url: String
+)
