+++ member-service/src/main/java/com/hedvig/memberservice/services/trustpilot/TrustpilotReviewService.kt
@@ -10,8 +10,3 @@ interface TrustpilotReviewService {
         locale: Locale?
     ): TrustpilotReviewInvitation?
 }
-
-data class TrustpilotReviewInvitation(
-    val id: String,
-    val url: String
-)
